--- airbnb/src/Contexts/Searching/context.ts
@@ -1,3 +1,0 @@
-// createContext
-// Provider
-// 이니셜 상태
