+++ airbnb/src/index.tsx
@@ -2,6 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from '@/App';
 
+if (process.env.NODE_ENV === 'development') {
+  const { worker } = require('./mocks/browser');
+  worker.start();
+}
+
+fetch('/').then(console.log);
+
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
 );
