+++ airbnb/src/types/Theme.d.ts
@@ -1,4 +1,3 @@
-import React from 'react';
 // 도대체 얘가 있으면 왜 되는건지 물어봐야함 이건 진짜 학회 논의 감
 
 // TODO: 발표때 모듈 타입 확장 시킨 과정 발표할것
